requirements.txt
boto3>=1.34
botocore>=1.34
google-api-core>=2.0
google-api-python-client>=2.0
google-auth>=2.0
google-auth-httplib2>=0.1
google-auth-oauthlib>=1.0
googleapis-common-protos>=1.50
httpx>=0.24
num2words>=0.5
openai>=1.0
proto-plus>=1.20
protobuf>=4.0
python-dotenv>=1.0.0
requests>=2.28
requests-oauthlib>=1.3
ytmusicapi>=1.12.0
tweepy>=4.0
uritemplate>=4.0
